Add optional installation of static library.
I tried to make it so that "make install" does not depend on "static" (to avoid compiling the static library when this wasn't really asked), by excluding some bits from "static". This means to install the static library (if compiled), it must be explicitly required with "cmake --install /path/to/build/dir --component static".

The main drawback is that this only works on Linux because on Windows the ARCHIVE component (which contains the static library) also contains the DLL export files (which is needed if other code wants to link with gstlearn).