Remove installation of NLopt when installing the static target.
The previous version was assuming that NLopt had been built in a specific location but it can come from any source (e.g., a system package). Moreover, as it is just an external dependency, installing it as part of this package is weird, so it's best to leave it as an external thing.

Also updated some comments about dependencies (gmtsph and NLopt).