Make dependency on NLopt and gmtsph always PRIVATE, even for gstlearn::static.
The static library still depends on those but this is actually correctly handled by CMake (in the installed package, the INTERFACE_LINK_LIBRARIES uses $<LINK_ONLY:...> for those so that there is a dependency on the actual libraries but without pulling in the includes, defines etc.). The previous commit that made the dependency PUBLIC was therefore wrong.

Also updated a related comment in install.cmake.